+++ r5dev/public/tier1/utlrbtree.h
@@ -151,11 +151,11 @@ public:
 	// Left at growSize = 0, the memory will first allocate 1 element and double in size
 	// at each increment.
 	// LessFunc_t is required, but may be set after the constructor using SetLessFunc() below
-	CUtlRBTree(int64 growSize = 0, int64 initSize = 0, const LessFunc_t& lessfunc = 0);
+	CUtlRBTree(ssize_t growSize = 0, ssize_t initSize = 0, const LessFunc_t& lessfunc = 0);
 	CUtlRBTree(const LessFunc_t& lessfunc);
 	~CUtlRBTree();
 
-	void EnsureCapacity(int64 num);
+	void EnsureCapacity(ssize_t num);
 
 	// NOTE: CopyFrom is fast but dangerous! It just memcpy's all nodes - it does NOT run copy constructors, so
 	//       it is not a true deep copy (i.e 'T' must be POD for this to work - e.g CUtlString will not work).
@@ -215,7 +215,7 @@ public:
 	// NOTE: the returned 'index' will be valid as long as the element remains in the tree
 	//       (other elements being added/removed will not affect it)
 	I  Insert(T const& insert);
-	void Insert(const T* pArray, int64 nItems);
+	void Insert(const T* pArray, ssize_t nItems);
 	I  InsertIfNotFound(T const& insert);
 
 	// Find method
@@ -384,7 +384,7 @@ protected:
 //-----------------------------------------------------------------------------
 
 template < class T, class I, typename L, class M >
-inline CUtlRBTree<T, I, L, M>::CUtlRBTree(int64 growSize, int64 initSize, const LessFunc_t& lessfunc) :
+inline CUtlRBTree<T, I, L, M>::CUtlRBTree(ssize_t growSize, ssize_t initSize, const LessFunc_t& lessfunc) :
 	m_LessFunc(lessfunc),
 	m_Elements(growSize, initSize),
 	m_Root(InvalidIndex()),
@@ -397,7 +397,7 @@ inline CUtlRBTree<T, I, L, M>::CUtlRBTree(int64 growSize, int64 initSize, const 
 
 template < class T, class I, typename L, class M >
 inline CUtlRBTree<T, I, L, M>::CUtlRBTree(const LessFunc_t& lessfunc) :
-	m_Elements((int64)0, (int64)0),
+	m_Elements((ssize_t)0, (ssize_t)0),
 	m_LessFunc(lessfunc),
 	m_Root(InvalidIndex()),
 	m_NumElements(0),
@@ -414,7 +414,7 @@ inline CUtlRBTree<T, I, L, M>::~CUtlRBTree()
 }
 
 template < class T, class I, typename L, class M >
-inline void CUtlRBTree<T, I, L, M>::EnsureCapacity(int64 num)
+inline void CUtlRBTree<T, I, L, M>::EnsureCapacity(ssize_t num)
 {
 	m_Elements.EnsureCapacity(num);
 }
@@ -1509,7 +1509,7 @@ I CUtlRBTree<T, I, L, M>::Insert(T const& insert)
 
 
 template < class T, class I, typename L, class M >
-void CUtlRBTree<T, I, L, M>::Insert(const T* pArray, int64 nItems)
+void CUtlRBTree<T, I, L, M>::Insert(const T* pArray, ssize_t nItems)
 {
 	while (nItems--)
 	{
